Add a PodLogger class
